Refactoring and fixes
