+++ consts/contract.ts
@@ -1,0 +1,8 @@
+
+
+export const CONTRACT_ADDRESS = '0x000a9a1bf96abc37d4959f395f74d0b00d61ce716dab42789fc2a68249655780'
+
+
+
+// mainet
+// export const CONTRACT_ADDRESS = '0x0168d11d36cd342d52d077296922e7cd505060778464c2d28bc28b4f3a9dc41c';
