--- utils/connect.ts
@@ -1,6 +1,0 @@
-import { connect } from "starknetkit";
-
-export const connectWallet = async() => {
-	const { wallet } = await connect();
-	return wallet
-}
